Se modulariza el proyecto
