+++ index.html
@@ -65,7 +65,7 @@
         </div> <!--.row-->
     </div><!--.container-->
 
-    <script src="js/app.js"></script>
+    <script src="js/app.js" type="module"></script>
 
 </body>
 </html>
