+++ js/selectors.js
@@ -1,0 +1,10 @@
+export const petInput = document.querySelector('#mascota');
+export const ownerInput = document.querySelector('#propietario');
+export const phoneInput = document.querySelector('#telefono');
+export const dateInput = document.querySelector('#fecha');
+export const hourInput = document.querySelector('#hora');
+export const symptomInput = document.querySelector('#sintomas');
+
+//UI
+export const form = document.querySelector('#nueva-cita');
+export const quotesContainer = document.querySelector('#citas');
